fix: .env error and scrapper timeout error (#6)
* fix: some errors when loading env

* chore: add some debug

* chore at debug

* refactor: information collect for Linkedin without need of credentials

* fix: handle absence of gesture for empty results and error handling

* refactor: response format for social networks

* fix: image dl error and add error handling

* feat: add code error

+++ src/api/api.py
@@ -77,11 +77,12 @@ async def search_by_face(upload_file: UploadFile, query: str):
     res = faces_compare(dir_name, fp, query)
 
     if type(res) is str:
-        HTTPException(
+        raise HTTPException(
             400,
-            detail={"data": res}
+            detail=res
         )
 
-    os.remove(fp)
+    if os.path.isfile(fp):
+        os.remove(fp)
 
     return res
